- Disable verification tasks for decomposition and diagonalization.
- Computation for the code in issue #1071 has decreased by another
10-20% or so.

The cumulative savings of this change, in combination with the previous
fixes has decreased conversion from 2.7 seconds to 1.25 seconds and
optimization from 12.6 seconds to 5.4 seconds.